services/chunking-service/requirements.txt
# Core dependencies
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Azure services
openai==1.54.0
httpx==0.27.0

# Storage
minio==7.2.3
psycopg2-binary==2.9.9
qdrant-client==1.7.0

# Message queue
pika==1.3.2

# Text processing
tiktoken==0.5.2
nltk==3.8.1

# Utilities
python-json-logger==2.0.7
